Fixed test class for reservation
